+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## Changelog
 
+### 12.53.0
+
+* `[strutil]` Added method `B` for choosing value by condition
+* `[system/process]` Tests updated for compatibility with GitHub Actions
+
 ### 12.52.0
 
 * `[fmtc]` Added method `If` for conditional message printing